Deployment/requirements.txt
waitress
numpy
Pillow
streamlit
tensorflow
tensorflow_hub
joblib